Made fight a little more interesting

+++ src/ch/ilikechickenwings/TXTRAP/Entity/Damageable.java
@@ -2,7 +2,7 @@ package ch.ilikechickenwings.TXTRAP.Entity;
 
 public interface Damageable {
 
-	public void getDamaged(float damage);
+	public void getDamaged(double damage);
